// src/event-calendar/calendar/components/header/calendar-header.tsx
import {
	Columns,
	Grid3x3,
	List,
	Plus,
	Grid2x2,
	CalendarRange,
} from "lucide-react"

import { Button } from "@fitodac/shadcn/button"

import { UserSelect } from "./user-select"
import { TodayButton } from "./today-button"
import { DateNavigator } from "./date-navigator"
import { AddEventDialog } from "../dialogs/add-event-dialog"
import { useCalendar } from "../../contexts/calendar-context"

import type { IEvent } from "../../interfaces"
import type { TCalendarView, TNavBar } from "../../types"

interface IProps {
	view: TCalendarView
	events: IEvent[]
	navBar?: TNavBar
}

export function CalendarHeader({ view, events, navBar = {} }: IProps) {
	const { LinkComponent } = useCalendar()

	return (
		<div className="flex flex-col gap-4 border-b p-4 lg:flex-row lg:items-center lg:justify-between">
			<div className="flex items-center gap-3">
				<TodayButton />
				<DateNavigator view={view} events={events} />
			</div>

			<div className="flex flex-col items-center gap-1.5 sm:flex-row sm:justify-between">
				<div className="flex w-full items-center gap-1.5">
					<div className="inline-flex first:rounded-r-none last:rounded-l-none [&:not(:first-child):not(:last-child)]:rounded-none">
						{navBar.day && (
							<Button
								asChild
								aria-label="View by day"
								size="icon"
								variant={view === "day" ? "default" : "outline"}
								className="rounded-r-none [&_svg]:size-5"
							>
								<LinkComponent href={navBar.day}>
									<List strokeWidth={1.8} />
								</LinkComponent>
							</Button>
						)}

						{navBar.week && (
							<Button
								asChild
								aria-label="View by week"
								size="icon"
								variant={view === "week" ? "default" : "outline"}
								className="-ml-px rounded-none [&_svg]:size-5"
							>
								<LinkComponent href={navBar.week}>
									<Columns strokeWidth={1.8} />
								</LinkComponent>
							</Button>
						)}

						{navBar.month && (
							<Button
								asChild
								aria-label="View by month"
								size="icon"
								variant={view === "month" ? "default" : "outline"}
								className="-ml-px rounded-none [&_svg]:size-5"
							>
								<LinkComponent href={navBar.month}>
									<Grid2x2 strokeWidth={1.8} />
								</LinkComponent>
							</Button>
						)}

						{navBar.year && (
							<Button
								asChild
								aria-label="View by year"
								size="icon"
								variant={view === "year" ? "default" : "outline"}
								className="-ml-px rounded-none [&_svg]:size-5"
							>
								<LinkComponent href={navBar.year}>
									<Grid3x3 strokeWidth={1.8} />
								</LinkComponent>
							</Button>
						)}

						{navBar.agenda && (
							<Button
								asChild
								aria-label="View by agenda"
								size="icon"
								variant={view === "agenda" ? "default" : "outline"}
								className="-ml-px rounded-l-none [&_svg]:size-5"
							>
								<LinkComponent href={navBar.agenda}>
									<CalendarRange strokeWidth={1.8} />
								</LinkComponent>
							</Button>
						)}
					</div>

					<UserSelect />
				</div>

				<AddEventDialog>
					<Button className="w-full sm:w-auto">
						<Plus />
						Add Event
					</Button>
				</AddEventDialog>
			</div>
		</div>
	)
}
